+++ src/main/kotlin/com/github/strindberg/emacsj/actions/kill/CopyAction.kt
@@ -1,0 +1,7 @@
+package com.github.strindberg.emacsj.actions.kill
+
+import com.github.strindberg.emacsj.kill.EditHandler
+import com.github.strindberg.emacsj.kill.Type
+import com.intellij.openapi.editor.actions.TextComponentEditorAction
+
+class CopyAction : TextComponentEditorAction(EditHandler(Type.COPY))