+++ src/main/kotlin/com/github/strindberg/emacsj/actions/kill/KillWholeLineAction.kt
@@ -4,4 +4,4 @@ import com.github.strindberg.emacsj.kill.KillLineHandler
 import com.github.strindberg.emacsj.kill.KillType
 import com.intellij.openapi.editor.actions.TextComponentEditorAction
 
-class KillLineAction : TextComponentEditorAction(KillLineHandler(KillType.REST_OF_LINE))
+class KillWholeLineAction : TextComponentEditorAction(KillLineHandler(KillType.WHOLE_LINE))