+++ src/main/kotlin/com/github/strindberg/emacsj/kill/AppendNextKillHandler.kt
@@ -1,0 +1,13 @@
+package com.github.strindberg.emacsj.kill
+
+import com.intellij.openapi.actionSystem.DataContext
+import com.intellij.openapi.editor.Caret
+import com.intellij.openapi.editor.Editor
+import com.intellij.openapi.editor.actionSystem.EditorActionHandler
+
+class AppendNextKillHandler : EditorActionHandler() {
+
+    override fun doExecute(editor: Editor, caret: Caret?, dataContext: DataContext) {
+        // Do nothing
+    }
+}