+++ src/main/kotlin/com/github/strindberg/emacsj/kill/EditHandler.kt
@@ -1,0 +1,22 @@
+package com.github.strindberg.emacsj.kill
+
+import com.intellij.openapi.actionSystem.DataContext
+import com.intellij.openapi.editor.Caret
+import com.intellij.openapi.editor.Editor
+import com.intellij.openapi.editor.actionSystem.EditorWriteActionHandler
+
+class EditHandler(val type: Type) : EditorWriteActionHandler() {
+
+    override fun executeWriteAction(editor: Editor, caret: Caret?, dataContext: DataContext) {
+        val primary = caret ?: editor.caretModel.primaryCaret
+        if (editor.selectionModel.hasSelection()) {
+            KillUtil.cutOrCopy(
+                type = type,
+                editor = editor,
+                textStartOffset = editor.selectionModel.selectionStart,
+                textEndOffset = editor.selectionModel.selectionEnd,
+                primary.offset == editor.selectionModel.selectionStart,
+            )
+        }
+    }
+}