+++ src/main/kotlin/com/github/strindberg/emacsj/zap/ZapDelegate.kt
@@ -1,12 +1,11 @@
 package com.github.strindberg.emacsj.zap
 
-import java.awt.datatransfer.StringSelection
 import java.awt.event.KeyEvent
 import java.util.UUID
+import com.github.strindberg.emacsj.kill.KillUtil
 import com.github.strindberg.emacsj.search.CommonUI
 import com.github.strindberg.emacsj.search.RestorableActionHandler
 import com.github.strindberg.emacsj.search.RestorableTypedActionHandler
-import com.github.strindberg.emacsj.word.substring
 import com.github.strindberg.emacsj.word.text
 import com.github.strindberg.emacsj.zap.ZapType.BACKWARD_TO
 import com.github.strindberg.emacsj.zap.ZapType.BACKWARD_UP_TO
@@ -20,7 +19,6 @@ import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.editor.actionSystem.EditorAction
 import com.intellij.openapi.editor.actionSystem.EditorActionManager
 import com.intellij.openapi.editor.actionSystem.TypedAction
-import com.intellij.openapi.ide.CopyPasteManager
 import org.jetbrains.annotations.VisibleForTesting
 
 class ZapDelegate(val editor: Editor, val type: ZapType) {
@@ -54,8 +52,9 @@ class ZapDelegate(val editor: Editor, val type: ZapType) {
                                 BACKWARD_TO, BACKWARD_UP_TO -> Pair(previousCharacter(editor.text, caret.offset, charTyped), caret.offset)
                             }
                             WriteCommandAction.runWriteCommandAction(editor.project, "Zap ${type.name.lowercase()}", undoGroupId, {
-                                CopyPasteManager.getInstance().setContents(StringSelection(document.substring(start, end)))
-                                document.deleteString(start, end)
+                                if (start != end) {
+                                    KillUtil.cut(editor, start, end, type in listOf(BACKWARD_TO, BACKWARD_UP_TO))
+                                }
                             })
                         }
                         cancel()